Add experimental njit features (#15)
Add experimental njit features

The features are not used in the main loop, but are added as starting points for 
future improvements. The idea is to use SharedMemory as a cache for forcing
chunks, and njit/numba for interpolation. There should be one process for
reading forcing, one process for computing interpolation and one process for
writing. Multiprocessing is required rather than multithreading to avoid GIL lock
issues.